Arreglo zoom

+++ trunk/src/codigo/TP taller/TP taller/Vista/Dibujables/Dibujable.h
@@ -23,7 +23,7 @@ public:
 	Dibujable();
 	~Dibujable();
 	virtual void setColor(int* rgb, int a);
-	virtual void dibujar(SDL_Renderer* renderer, int corrimiento, int escalaZoom, int posZoomX, int posZoomY);
+	virtual void dibujar(SDL_Renderer* renderer, int corrimiento, float escalaZoom, int posZoomX, int posZoomY);
 	virtual DibujableSerializado getDibujableSerializado(int& tamano);
 };
