+++ trunk/src/codigo/TP taller/TP taller/Vista/Sprite/Sprite.cpp
@@ -33,7 +33,7 @@ Sprite::~Sprite(){
 	delete []this->recCuadro;
 }
 
-void Sprite::dibujar(SDL_Renderer *renderer, int corrimiento, int escalaZoom, int posZoomX, int posZoomY){
+void Sprite::dibujar(SDL_Renderer *renderer, int corrimiento, float escalaZoom, int posZoomX, int posZoomY){
 
 	SDL_Rect rect = this->recDest;
